requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==1.4.46
PyPDF2==3.0.1
python-docx==1.1.0
chromadb==0.4.15
sentence-transformers==2.2.2
openai==1.3.0
langchain==0.0.335
langchain-openai==0.0.2
tiktoken==0.11.0
